alert_service/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.11
botocore==1.34.11
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
confluent-kafka==2.3.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iso3166==2.1.1
jmespath==1.0.1
kafka-python==2.0.2
leaflet==0.0.3
mapbox==0.18.1
msgpack==1.0.7
packaging==23.2
pika==1.3.2
pillow==10.2.0
polyline==2.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
zope.interface==6.1
